Product service unit tests

+++ backend/src/main/java/com/pedrosa/dscatalog/services/CategoryService.java
@@ -62,7 +62,10 @@ public class CategoryService {
     
     @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-
+    	
+    	if (!repository.existsById(id)) {
+    		throw new ResourceNotFoundException("Id not found " + id);
+    	}
         try {
             repository.deleteById(id);
         } catch (DataIntegrityViolationException e) {
